Detect loops in parent thread state
Detect loops in parent thread state by noting the starting point of the
search and aborting whenever the current thread pointer is equal to the
place at which we started.

This prevents infinite loops like observed in
https://github.com/draios/falco/issues/208.

This doesn't address the underlying cause of what caused the thread
state to get corrupted in the first place. That's tracked by a separate
issue https://github.com/draios/sysdig/issues/752.